test: Move two helper functions out to VerifyFunctionFieldSamples.ts
To prepare for writing more tests for other fields.
