+++ tests/Query/Filter/ReferenceDocs/FilterReference/TagsFieldDocs.test.ts
@@ -1,42 +1,8 @@
-import { verifyAll } from 'approvals/lib/Providers/Jest/JestApprovals';
-import type { Task } from '../../../../../src/Task';
-import { FunctionField } from '../../../../../src/Query/Filter/FunctionField';
-import { groupHeadingsForTask } from '../../../../CustomMatchers/CustomMatchersForGrouping';
 import { SampleTasks } from '../../../../TestHelpers';
-import { verifyMarkdownForDocs } from '../../../../TestingTools/VerifyMarkdownTable';
-
-function verifyFunctionFieldGrouperSamplesOnTasks(customGroups: string[][], tasks: Task[]) {
-    verifyAll('Results of custom groupers', customGroups, (group) => {
-        const instruction = group[0];
-        const comment = group[1];
-        const grouper = new FunctionField().createGrouperFromLine(instruction);
-        const headings = groupHeadingsForTask(grouper!, tasks);
-        return `
-${instruction}
-${comment}
-=>
-${headings.join('\n')}
-====================================================================================
-`;
-    });
-}
-
-function verifyFunctionFieldGrouperSamplesForDocs(customGroups: string[][]) {
-    let markdown = '';
-    for (const group of customGroups) {
-        const instruction = group[0];
-        const comment = group[1];
-        markdown += `
-~~~text
-${instruction}
-~~~
-
-- ${comment}.
-
-`;
-    }
-    verifyMarkdownForDocs(markdown);
-}
+import {
+    verifyFunctionFieldGrouperSamplesForDocs,
+    verifyFunctionFieldGrouperSamplesOnTasks,
+} from './VerifyFunctionFieldSamples';
 
 describe('custom grouping by tag', () => {
     const customGroups = [
